splitting concatenation in submodules.
Not working: Full of circular imports
